25/6/2025

+++ src/main/java/com/ayatie/ayat/repository/QuraanRepository.java
@@ -3,5 +3,5 @@ package com.ayatie.ayat.repository;
 import com.ayatie.ayat.model.Quraan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuraanRepository extends JpaRepository<Quraan, Long> {
+public interface QuraanRepository extends JpaRepository<Quraan, Integer> {
 }
